Refactored code to make sure that we have different generators for the multichannel mip data and the normal mip data
